chore(snappixel): move sdk loader to a separate file
